Upgrade to asp.net 2.1 and EF 2.1. Swap out the oracle mysql EF provider with Pomelo, and refactor a few things to make that work

+++ TASVideos.Legacy/Imports/PublicationRatingImporter.cs
@@ -43,7 +43,7 @@ namespace TASVideos.Legacy.Imports
 					UserId = u.Id,
 					PublicationId = r.PublicationId,
 					Type = r.Type,
-					Value = r.Value
+					Value = (double)r.Value
 				})
 				.ToList();
